Add prayer to Jupiter.
This is from the Iguvine Tables[0].  Whoever was writing that section of the
Wikipedia article decided to cut off in the middle of a sentence, so I had to
improvise an ending.  But hey, I'm butchering an ancient prayer to provide some
small amount of amusement, so I might as well start making stuff up, right?

[0]: http://en.wikipedia.org/wiki/Iguvine_Tables#Text_sample